+++ src/ScottPlot5/ScottPlot5/PlotStyles/Light.cs
@@ -1,0 +1,16 @@
+﻿namespace ScottPlot.PlotStyles;
+
+public class Light : PlotStyle
+{
+    public Light()
+    {
+        Palette = new Palettes.Category10();
+        Axes = Colors.Black;
+        GridMajorLine = Colors.Black.WithOpacity(.1);
+        FigureBackground = Colors.White;
+        DataBackGround = Colors.White;
+        LegendBackground = Colors.White;
+        LegendFont = Colors.Black;
+        LegendOutline = Colors.Black;
+    }
+}